Collect only OK,(200,302) pages
The Archiver collects *every* page, means it tries to collect all 404 pages also. 
The spider.every_ok_page ensures, the Archiver does not collect 404 pages. 
I have tested and it collects 200 (OK) and also 302 (Redirected) pages. 
As i use this very extensively, the 404 with invalid urls had stacked up for some sites like /typo3/.../typo3/.../typo3/ 
and so the 404 collection had slowed down the actual archiving task of this script 
or even render the archiving of a whole page unsuccessful.